+++ src/LoggerModem.h
@@ -43,7 +43,7 @@
  * Variable objects to be tied to a loggerModem.  These are measured by a modem,
  * but are implemented as calculated variables.
  *
- * @note  The modem is NOT set up as a sensor.  ALl of these variables for the
+ * @note  The modem is NOT set up as a sensor.  ALL of these variables for the
  * modem object are actually being called as calculated variables where the
  * calculation function is to ask the modem object for the values from the last
  * time it connected to the internet.
@@ -68,6 +68,9 @@
  * RSSI is a rough calculation, so it has 0 decimal place resolution
  */
 #define MODEM_RSSI_RESOLUTION 0
+/// @brief The bit mask for loggerModem::_pollModemMetaData to enable RSSI
+/// polling.
+#define MODEM_RSSI_ENABLE_BITMASK 0b00000001
 /// @brief Variable name in
 /// [ODM2 controlled vocabulary](http://vocabulary.odm2.org/variablename/);
 /// "RSSI"
@@ -94,6 +97,9 @@
  * Percent signal is a rough calculation, so it has 0 decimal place resolution
  */
 #define MODEM_PERCENT_SIGNAL_RESOLUTION 0
+/// @brief The bit mask for loggerModem::_pollModemMetaData to enable percent
+/// signal polling.
+#define MODEM_PERCENT_SIGNAL_ENABLE_BITMASK 0b00000010
 /// @brief Variable name in
 /// [ODM2 controlled vocabulary](http://vocabulary.odm2.org/variablename/);
 /// "signalPercent"
@@ -124,6 +130,9 @@
  * Battery state is a code value; it has 0 decimal place resolution
  */
 #define MODEM_BATTERY_STATE_RESOLUTION 0
+/// @brief The bit mask for loggerModem::_pollModemMetaData to enable modem
+/// battery charging state polling.
+#define MODEM_BATTERY_STATE_ENABLE_BITMASK 0b00000100
 /// @brief Variable name in
 /// [ODM2 controlled vocabulary](http://vocabulary.odm2.org/variablename/);
 /// "batteryChargeState"
@@ -151,6 +160,9 @@
 /// @brief Decimals places in string representation; battery charge percent
 /// should have 0.
 #define MODEM_BATTERY_PERCENT_RESOLUTION 0
+/// @brief The bit mask for loggerModem::_pollModemMetaData to enable modem
+/// battery percent polling.
+#define MODEM_BATTERY_PERCENT_ENABLE_BITMASK 0b00001000
 /// @brief Variable name in
 /// [ODM2 controlled vocabulary](http://vocabulary.odm2.org/variablename/);
 /// "batteryVoltage"
@@ -181,6 +193,9 @@
  * No supported module has higher than 1mV resolution in battery reading.
  */
 #define MODEM_BATTERY_VOLTAGE_RESOLUTION 0
+/// @brief The bit mask for loggerModem::_pollModemMetaData to enable modem
+/// battery voltage polling.
+#define MODEM_BATTERY_VOLTAGE_ENABLE_BITMASK 0b00010000
 /// @brief Variable name in
 /// [ODM2 controlled vocabulary](http://vocabulary.odm2.org/variablename/);
 /// "batteryVoltage"
@@ -210,6 +225,9 @@
  * Most modules that can measure temperature measure to 0.1°C
  */
 #define MODEM_TEMPERATURE_RESOLUTION 1
+/// @brief The bit mask for loggerModem::_pollModemMetaData to enable modem
+/// temperature polling.
+#define MODEM_TEMPERATURE_ENABLE_BITMASK 0b00100000
 /// @brief Variable name in
 /// [ODM2 controlled vocabulary](http://vocabulary.odm2.org/variablename/);
 /// "temperature"
@@ -236,6 +254,9 @@
 /// @brief Decimals places in string representation; total active time should
 /// have 3.
 #define MODEM_ACTIVATION_RESOLUTION 3
+/// @brief The bit mask for loggerModem::_pollModemMetaData to enable modem
+/// activation time polling.
+#define MODEM_ACTIVATION_ENABLE_BITMASK 0b01000000
 /// @brief Variable name in
 /// [ODM2 controlled vocabulary](http://vocabulary.odm2.org/variablename/);
 /// "timeElapsed"
@@ -260,6 +281,9 @@
 /// @brief Decimals places in string representation; total powered time should
 /// have 3.
 #define MODEM_POWERED_RESOLUTION 3
+/// @brief The bit mask for loggerModem::_pollModemMetaData to enable modem
+/// power time polling
+#define MODEM_POWERED_ENABLE_BITMASK 0b10000000
 /// @brief Variable name in
 /// [ODM2 controlled vocabulary](http://vocabulary.odm2.org/variablename/);
 /// "timeElapsed"
@@ -339,6 +363,16 @@ class loggerModem {
      * @return **String** The modem name
      */
     String getModemName(void);
+
+    /**
+     * @brief Get a detailed printable description of the modem.
+     *
+     * @note These values are polled for and cached in memory till needed
+     *
+     * @return **String** The concatenated name, hardware version, firmware
+     * version, and serial number fo the modem..
+     */
+    String getModemDevId(void);
 
     /**
      * @brief Set up the modem before first use.
@@ -535,6 +569,7 @@ class loggerModem {
      * called after the modem is connected to the internet.
      */
     /**@{*/
+
     /**
      * @brief Query the modem for the current signal quality and write the
      * results to the supplied non-constant references.
@@ -572,13 +607,45 @@ class loggerModem {
      */
     virtual float getModemChipTemperature(void) = 0;
 
+
+    /**
+     * @brief Enables metadata polling for one or more modem measured
+     * variables. Setting this to 0b11111111 will enable polling for all modem
+     * measured variables.
+     *
+     * @note This will **not** disable polling for any unset bits in the
+     * provided bitmask.  It will only enable those bits that are set.
+     */
+    void enableMetadataPolling(uint8_t pollingBitmask);
+
+    /**
+     * @brief Disables metadata polling for one or more modem measured
+     * variables.  Setting this to 0b11111111 will disable polling for all modem
+     * measured variables.
+     *
+     * @note This will **not** enable polling for any unset bits in the
+     * provided bitmask.  It will only disable polling for those bits that are
+     * set.
+     */
+    void disableMetadataPolling(uint8_t pollingBitmask);
+
+    /**
+     * @brief Sets the bitmask for modem metadata polling.
+     *
+     * This will enable polling for 1 bits and disable polling for 0 bits.
+     * Setting this to 0 (0b00000000) will disable polling for all metadata
+     * parameters.  Setting it to 256 (0b11111111) will enable polling for all
+     * parameters.
+     */
+    void setMetadataPolling(uint8_t pollingBitmask);
+
     /**
      * @brief Query the modem for signal quality, battery, and temperature
      * information and store the values to the static internal variables.
      *
-     * @return **bool** True indicates that the communication with the modem was
-     * successful and the values of the internal static variables should be
-     * valid.
+     * @return **bool** True indicates that the communication with the modem
+     * was successful and the values of the internal static variables should
+     * be valid.
      */
     virtual bool updateModemMetadata(void);
     /**@}*/
@@ -989,6 +1056,48 @@ class loggerModem {
 
     // modemType gsmModem;
     // modemClientType gsmClient;
+
+    /**
+     * @brief The modem hardware version.
+     *
+     * Set in #modemSetup().
+     * Returned as a portion of the #getModemDevId().
+     */
+    String _modemHwVersion;
+    /**
+     * @brief The modem firmware version.
+     *
+     * Set in #modemSetup().
+     * Returned as a portion of the #getModemDevId().
+     */
+    String _modemFwVersion;
+    /**
+     * @brief The modem serial number.
+     *
+     * Set in #modemSetup().
+     * Returned as a portion of the #getModemDevId().
+     */
+    String _modemSerialNumber;
+
+    /**
+     * @brief An 8-bit code for the enabled modem polling variables
+     *
+     * Setting a bit to 0 will disable polling, to 1 will enable it.  By default
+     * no polling is enabled to save time and power by not requesting
+     * unnecessary information from the modem.  When modem measured variables
+     * are attached to a modem, polling for those results is automatically
+     * enabled.
+     *
+     * Bit | Variable Class | Relevent Define
+     * ----|----------------|----------------
+     *  0  | #Modem_RSSI | #MODEM_RSSI_ENABLE_BITMASK
+     *  1  | #Modem_SignalPercent | #MODEM_PERCENT_SIGNAL_ENABLE_BITMASK
+     *  2  | #Modem_BatteryState | #MODEM_BATTERY_STATE_ENABLE_BITMASK
+     *  3  | #Modem_BatteryPercent | #MODEM_BATTERY_PERCENT_ENABLE_BITMASK
+     *  4  | #Modem_BatteryVoltage | #MODEM_BATTERY_VOLTAGE_ENABLE_BITMASK
+     *  5  | #Modem_Temp | #MODEM_TEMPERATURE_ENABLE_BITMASK
+     */
+    uint8_t _pollModemMetaData = 0;
 };
 
 // typedef float (loggerModem::_*loggerGetValueFxn)(void);
@@ -1017,7 +1126,9 @@ class Modem_RSSI : public Variable {
                         const char* varCode = MODEM_RSSI_DEFAULT_CODE)
         : Variable(&parentModem->getModemRSSI, (uint8_t)MODEM_RSSI_RESOLUTION,
                    &*MODEM_RSSI_VAR_NAME, &*MODEM_RSSI_UNIT_NAME, varCode,
-                   uuid) {}
+                   uuid) {
+        parentModem->enableMetadataPolling(MODEM_RSSI_ENABLE_BITMASK);
+    }
     /**
      * @brief Destroy the Modem_RSSI object - no action needed.
      */
@@ -1050,7 +1161,9 @@ class Modem_SignalPercent : public Variable {
         : Variable(&parentModem->getModemSignalPercent,
                    (uint8_t)MODEM_PERCENT_SIGNAL_RESOLUTION,
                    &*MODEM_PERCENT_SIGNAL_VAR_NAME,
-                   &*MODEM_PERCENT_SIGNAL_UNIT_NAME, varCode, uuid) {}
+                   &*MODEM_PERCENT_SIGNAL_UNIT_NAME, varCode, uuid) {
+        parentModem->enableMetadataPolling(MODEM_PERCENT_SIGNAL_ENABLE_BITMASK);
+    }
     /**
      * @brief Destroy the Modem_SignalPercent object - no action needed.
      */
@@ -1086,7 +1199,9 @@ class Modem_BatteryState : public Variable {
         : Variable(&parentModem->getModemBatteryChargeState,
                    (uint8_t)MODEM_BATTERY_STATE_RESOLUTION,
                    &*MODEM_BATTERY_STATE_VAR_NAME,
-                   &*MODEM_BATTERY_STATE_UNIT_NAME, varCode, uuid) {}
+                   &*MODEM_BATTERY_STATE_UNIT_NAME, varCode, uuid) {
+        parentModem->enableMetadataPolling(MODEM_BATTERY_STATE_ENABLE_BITMASK);
+    }
     /**
      * @brief Destroy the Modem_BatteryState object - no action needed.
      */
@@ -1122,7 +1237,10 @@ class Modem_BatteryPercent : public Variable {
         : Variable(&parentModem->getModemBatteryChargePercent,
                    (uint8_t)MODEM_BATTERY_PERCENT_RESOLUTION,
                    &*MODEM_BATTERY_PERCENT_VAR_NAME,
-                   &*MODEM_BATTERY_PERCENT_UNIT_NAME, varCode, uuid) {}
+                   &*MODEM_BATTERY_PERCENT_UNIT_NAME, varCode, uuid) {
+        parentModem->enableMetadataPolling(
+            MODEM_BATTERY_PERCENT_ENABLE_BITMASK);
+    }
     /**
      * @brief Destroy the Modem_BatteryPercent object - no action needed.
      */
@@ -1158,7 +1276,10 @@ class Modem_BatteryVoltage : public Variable {
         : Variable(&parentModem->getModemBatteryVoltage,
                    (uint8_t)MODEM_BATTERY_VOLTAGE_RESOLUTION,
                    &*MODEM_BATTERY_VOLTAGE_VAR_NAME,
-                   &*MODEM_BATTERY_VOLTAGE_UNIT_NAME, varCode, uuid) {}
+                   &*MODEM_BATTERY_VOLTAGE_UNIT_NAME, varCode, uuid) {
+        parentModem->enableMetadataPolling(
+            MODEM_BATTERY_VOLTAGE_ENABLE_BITMASK);
+    }
     /**
      * @brief Destroy the Modem_BatteryVoltage object - no action needed.
      */
@@ -1193,7 +1314,9 @@ class Modem_Temp : public Variable {
         : Variable(&parentModem->getModemTemperature,
                    (uint8_t)MODEM_TEMPERATURE_RESOLUTION,
                    &*MODEM_TEMPERATURE_VAR_NAME, &*MODEM_TEMPERATURE_UNIT_NAME,
-                   varCode, uuid) {}
+                   varCode, uuid) {
+        parentModem->enableMetadataPolling(MODEM_TEMPERATURE_ENABLE_BITMASK);
+    }
     /**
      * @brief Destroy the Modem_Temp object - no action needed.
      */