+++ src/ModularSensors.h
@@ -13,8 +13,12 @@
 
 /**
  * @brief The current library version number
+ * https://semver.org/
+ * Add hypen '-' and alpha number for a branches unique tracking number
+ * A pre-release version will always be indicated as slightly ahead of the
+ * EnviroDIY branch that it is based on.
  */
-#define MODULAR_SENSORS_VERSION "0.34.0"
+#define MODULAR_SENSORS_VERSION "0.35.0"
 
 // To get all of the base classes for ModularSensors, include LoggerBase.
 // NOTE:  Individual sensor definitions must be included separately.