+++ src/dataPublisherBase.h
@@ -32,11 +32,10 @@
  * @def MS_SEND_BUFFER_SIZE
  * @brief Send Buffer
  *
- * This determines how many characters to set out at once over the TCP/UDP
- * connection.  Increasing this may decrease data use by a loger, while
- * decreasing it will save memory.  Do not make it smaller than 47 (to keep all
- * variable values with their UUID's) or bigger than 1500 (a typical TCP/UDP
- * Maximum Transmission Unit).
+ * This determines how many characters to set out at once over the TCP
+ * connection. Increasing this may decrease data use by a logger, while
+ * decreasing it will save memory. Do not make it smaller than 32 or bigger
+ * than 1500 (a typical TCP Maximum Transmission Unit).
  *
  * This can be changed by setting the build flag MS_SEND_BUFFER_SIZE when
  * compiling.
@@ -87,20 +86,10 @@ class dataPublisher {
      * logger.
      *
      * @param baseLogger The logger supplying the data to be published
-     * @param sendEveryX Currently unimplemented, intended for future use to
-     * enable caching and bulk publishing
-     * @param sendOffset Currently unimplemented, intended for future use to
-     * enable publishing data at a time slightly delayed from when it is
-     * collected
-     *
-     * @note It is possible (though very unlikey) that using this constructor
-     * could cause errors if the compiler attempts to initialize the publisher
-     * instance before the logger instance.  If you suspect you are seeing that
-     * issue, use the null constructor and a populated begin(...) within your
-     * set-up function.
+     * @param sendEveryX Interval (in units of the logging interval) between
+     * attempted data transmissions. Not respected by all publishers.
      */
-    explicit dataPublisher(Logger& baseLogger, uint8_t sendEveryX = 1,
-                           uint8_t sendOffset = 0);
+    explicit dataPublisher(Logger& baseLogger, int sendEveryX = 1);
     /**
      * @brief Construct a new data Publisher object.
      *
@@ -108,20 +97,10 @@ class dataPublisher {
      * @param inClient An Arduino client instance to use to print data to.
      * Allows the use of any type of client and multiple clients tied to a
      * single TinyGSM modem instance
-     * @param sendEveryX Currently unimplemented, intended for future use to
-     * enable caching and bulk publishing
-     * @param sendOffset Currently unimplemented, intended for future use to
-     * enable publishing data at a time slightly delayed from when it is
-     * collected
-     *
-     * @note It is possible (though very unlikey) that using this constructor
-     * could cause errors if the compiler attempts to initialize the publisher
-     * instance before the logger instance.  If you suspect you are seeing that
-     * issue, use the null constructor and a populated begin(...) within your
-     * set-up function.
+     * @param sendEveryX Interval (in units of the logging interval) between
+     * attempted data transmissions. Not respected by all publishers.
      */
-    dataPublisher(Logger& baseLogger, Client* inClient, uint8_t sendEveryX = 1,
-                  uint8_t sendOffset = 0);
+    dataPublisher(Logger& baseLogger, Client* inClient, int sendEveryX = 1);
     /**
      * @brief Destroy the data Publisher object - no action is taken.
      */
@@ -146,30 +125,19 @@ class dataPublisher {
      */
     void attachToLogger(Logger& baseLogger);
     /**
-     * @brief Set the parameters for frequency of sending and any offset, if
-     * needed.
+     * @brief Sets the interval (in units of the logging interval) between
+     * attempted data transmissions
      *
-     * @param sendEveryX Currently unimplemented, intended for future use to
-     * enable caching and bulk publishing
-     * @param sendOffset Currently unimplemented, intended for future use to
-     * enable publishing data at a time slightly delayed from when it is
-     * collected
-     *
-     * @note These parameters are not currently used!
+     * @param sendEveryX Interval (in units of the logging interval) between
+     * attempted data transmissions. Not respected by all publishers.
      */
-    void setSendFrequency(uint8_t sendEveryX, uint8_t sendOffset);
+    void setSendInterval(int sendEveryX);
 
     /**
      * @brief Begin the publisher - linking it to the client and logger.
      *
      * This can be used as an alternative to adding the logger and client in the
-     * constructor.  This is slightly "safer" because we expect the publishers
-     * to be created in the "global scope" and we cannot control the order in
-     * which objects in that global scope will be created.  That is, we cannot
-     * guarantee that the logger will actually be created before the publisher
-     * that wants to attach to it unless we wait to attach the publisher until
-     * in the setup or loop function of the main program.  In reality, it is
-     * very unlikely that this is necessary.
+     * constructor.
      *
      * @param baseLogger The logger supplying the data to be published
      * @param inClient An Arduino client instance to use to print data to.
@@ -187,13 +155,7 @@ class dataPublisher {
      * logger.
      *
      * This can be used as an alternative to adding the logger and client in the
-     * constructor.  This is slightly "safer" because we expect the publishers
-     * to be created in the "global scope" and we cannot control the order in
-     * which objects in that global scope will be created.  That is, we cannot
-     * guarantee that the logger will actually be created before the publisher
-     * that wants to attach to it unless we wait to attach the publisher until
-     * in the setup or loop function of the main program.  In reality, it is
-     * very unlikely that this is necessary.
+     * constructor.
      *
      * @param baseLogger The logger supplying the data to be published
      */
@@ -294,40 +256,54 @@ class dataPublisher {
     Client* _inClient = nullptr;
 
     /**
-     * @brief A buffer for outgoing data.
+     * @brief The buffer for outgoing data.
      */
     static char txBuffer[MS_SEND_BUFFER_SIZE];
     /**
-     * @brief Get the number of empty spots in the buffer.
-     *
-     * @return **int** The number of available characters in the buffer
+     * @brief The pointer to the client instance the TX buffer is writing to.
      */
-    static int bufferFree(void);
+    static Client* txBufferOutClient;
     /**
-     * @brief Fill the TX buffer with nulls ('\0').
+     * @brief The number of used characters in the TX buffer.
      */
-    static void emptyTxBuffer(void);
+    static size_t txBufferLen;
     /**
-     * @brief Write the TX buffer to a stream and also to the debugging
-     * port.
+     * @brief Initialize the TX buffer to be empty and start writing to the
+     * given client.
      *
-     * @param stream A pointer to an Arduino Stream instance to use to print
-     * data
-     * @param addNewLine True to add a new line character ("\n") at the end of
-     * the print
+     * @param client The client to transmit to.
      */
-    static void printTxBuffer(Stream* stream, bool addNewLine = false);
+    static void txBufferInit(Client* outClient);
+    /**
+     * @brief Append the given data to the TX buffer, flushing if necessary.
+     *
+     * @param data The data start pointer.
+     * @param length The number of bytes to add.
+     */
+    static void txBufferAppend(const char* data, size_t length);
+    /**
+     * @brief Append the given string to the TX buffer, flushing if necessary.
+     *
+     * @param s The null-terminated string to append.
+     */
+    static void txBufferAppend(const char* s);
+    /**
+     * @brief Append the given char to the TX buffer, flushing if necessary.
+     *
+     * @param c The char to append.
+     */
+    static void txBufferAppend(char c);
+    /**
+     * @brief Write the TX buffer contents to the initialized stream and also to
+     * the debugging port.
+     */
+    static void txBufferFlush();
 
     /**
-     * @brief Unimplemented; intended for future use to enable caching and bulk
-     * publishing.
+     * @brief Interval (in units of the logging interval) between
+     * attempted data transmissions. Not respected by all publishers.
      */
-    uint8_t _sendEveryX = 1;
-    /**
-     * @brief Unimplemented; intended for future use to enable publishing data
-     * at a time slightly delayed from when it is collected.
-     */
-    uint8_t _sendOffset = 0;
+    int _sendEveryX = 1;
 
     // Basic chunks of HTTP
     /**